pyproject.toml
[project]
name = "chat-x"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "grpcio>=1.69.0",
    "grpcio-tools>=1.69.0",
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "openai>=1.59.8",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
